Separate out problematic import so tests arent fucked
